Uint

+++ Sources/Wayland/Rect.swift
@@ -1,10 +1,10 @@
 public struct Rect {
 
-    public let dst_p0: (Float, Float)
-    public let dst_p1: (Float, Float)
+    public let dst_p0: (UInt, UInt)
+    public let dst_p1: (UInt, UInt)
     public let color: Color
 
-    public init(dst_p0: (Float, Float), dst_p1: (Float, Float), color: Color) {
+    public init(dst_p0: (UInt, UInt), dst_p1: (UInt, UInt), color: Color) {
         self.dst_p0 = dst_p0
         self.dst_p1 = dst_p1
         self.color = color
